Еще рефакторинг
